ZeekflowApi/requirements.txt
Flask==2.0.1
requests==2.25.1
scikit-learn==1.0.2
pandas==1.4.2
kafka-python==2.0.2